完成子回复功能

+++ app/Http/Controllers/RepliesController.php
@@ -24,12 +24,20 @@ class RepliesController extends Controller
      */
     public function store(ReplyRequest $request, Reply $reply): RedirectResponse
     {
+        $child = false;
+        if ($request->parent_id) {
+            $reply->parent_id = $request->parent_id;
+            $child = true;
+        }
+
         $reply->message = $request->message;
         $reply->user_id = Auth::id();
         $reply->topic_id = $request->topic_id;
         $reply->save();
 
-        return redirect()->to($reply->topic->slug . '#reply' . $reply->id)->with('success', 'Reply created successfully.');
+        return redirect()
+            ->to($reply->topic->slug . "?child=$child&reply_id=$reply->parent_id#reply$reply->id")
+            ->with('success', 'Reply created successfully.');
     }
 
     /**
@@ -41,9 +49,20 @@ class RepliesController extends Controller
      */
     public function destroy(Reply $reply): RedirectResponse
     {
+        // If the reply has children, do not delete it.
+        if ($reply->child()->exists()) {
+            return back()->with('danger', 'This reply has children, please delete them first.');
+        }
+
+        // For collapsed reply's child item.
+        $child = false;
+        if ($reply->parent_id) {
+            $child = true;
+        }
+
         $this->authorize('destroy', $reply);
         $reply->delete();
 
-        return redirect()->to($reply->topic->slug)->with('success', 'Deleted successfully.');
+        return redirect()->to($reply->topic->slug . "?child=$child&reply_id=$reply->parent_id#reply$reply->id")->with('success', 'Deleted successfully.');
     }
 }
